mcp-servers/pubmed-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pubmed-mcp-server"
version = "0.1.0"
description = "MCP server for searching PubMed/MEDLINE via NCBI E-utilities"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
pubmed-server = "pubmed_server:serve"

[tool.hatch.build.targets.wheel]
packages = ["src/pubmed_server"]
